requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
attrs==25.3.0
autocommand==2.2.2
backports.tarfile==1.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
connected-components-3d==3.23.0
contourpy==1.3.1
cycler==0.12.1
dijkstra3d==1.15.1
dill==0.3.9
docopt==0.6.2
edt==3.0.0
einops==0.8.1
fastremap==1.15.1
filelock==3.13.1
fill_voids==2.0.8
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.0.0
importlib_resources==6.5.2
inflect==7.3.1
iniconfig==2.1.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.4
joblib==1.4.2
kimimaro==5.0.0
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lungmask==0.2.20
Mako==1.3.9
MarkupSafe==2.1.5
matplotlib==3.10.1
monai==1.4.0
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.17
natsort==8.4.0
networkx==3.3
nibabel==5.3.2
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
osteoid==0.3.1
packaging==24.2
pandas==2.2.3
pathos==0.3.3
pillow==11.1.0
platformdirs==4.2.2
pluggy==1.5.0
pox==0.3.5
ppft==1.7.6.9
propcache==0.3.0
psutil==7.0.0
pycuda==2025.1
pydicom==3.0.1
pykwalify==1.8.0
PyMaxflow==1.3.2
pyparsing==3.2.2
pyradiomics==3.0.1
pytest==8.3.5
python-dateutil==2.9.0.post0
pytools==2025.1.1
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
requests==2.32.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
SimpleITK==2.4.1
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.3.13
tomli==2.0.1
torch==2.5.1+cu121
torch-geometric==2.6.1
tqdm==4.67.1
triton==3.1.0
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
xs3d==1.7.2
yarl==1.18.3
zipp==3.19.2
